% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AI_STAAR_Binary_SPA.R
\name{AI_STAAR_Binary_SPA}
\alias{AI_STAAR_Binary_SPA}
\title{Ancestry-Informed STAAR-SPA procedure using omnibus test}
\usage{
AI_STAAR_Binary_SPA(
  genotype,
  obj_nullmodel,
  annotation_phred = NULL,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2,
  rv_num_cutoff_max = 1e+09,
  tol = .Machine$double.eps^0.25,
  max_iter = 1000,
  SPA_p_filter = FALSE,
  p_filter_cutoff = 0.05,
  find_weight = FALSE
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants.}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glm}} function for unrelated samples or
\code{\link{fit_null_glmmkin}} function for related samples. Note that \code{\link{fit_null_glmmkin}}
is a wrapper of the \code{\link{glmmkin}} function from the \code{\link{GMMAT}} package.}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p).
Continuous scores should be given in PHRED score scale, where the PHRED score
of j-th variant is defined to be -10*log10(rank(-score_j)/total) across the genome. (Binary)
categorical scores should be taking values 0 or 1, where 1 is functional and 0 is
non-functional. If not provided, STAAR will perform the
Burden(1,25) and Burden(1,1) tests (default = NULL).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}

\item{rv_num_cutoff_max}{the cutoff of maximum number of variants of analyzing
a given variant-set (default = 1e+09).}

\item{tol}{a positive number specifying tolerance, the difference threshold for parameter
estimates in saddlepoint apporximation algorithm below which iterations should be stopped (default = ".Machine$double.eps^0.25").}

\item{max_iter}{a positive integers pecifying the maximum number of iterations for applying the saddlepoint approximation algorithm (default = "1000").}

\item{SPA_p_filter}{logical: are only the variants with a normal approximation based p-value smaller than a pre-specified threshold use the SPA method to recalculate the p-value, only used for imbalanced case-control setting (default = FALSE).}

\item{p_filter_cutoff}{threshold for the p-value recalculation using the SPA method, only used for imbalanced case-control setting (default = 0.05).}

\item{find_weight}{logical: should the ancestry group-specific weights and weighting scenario-specific p-values for each base test be saved as output (default = FALSE).}
}
\value{
A list with the following members:

\code{num_variant}: the number of variants with minor allele frequency > 0 and less than
\code{rare_maf_cutoff} in the given variant-set that are used for performing the
variant-set using STAAR.

\code{cMAC}: the cumulative minor allele count of variants with
minor allele frequency > 0 and less than \code{rare_maf_cutoff} in the given variant-set.

\code{RV_label}: the boolean vector indicating whether each variant in the given
variant-set has minor allele frequency > 0 and less than \code{rare_maf_cutoff}.

\code{results_STAAR_B}: the STAAR-B p-value that aggregated Burden(1,25) and Burden(1,1) together
with p-values of each test weighted by each annotation using Cauchy method.

\code{results_STAAR_B_1_25}: a vector of STAAR-B(1,25) p-values,
including Burden(1,25) p-value weighted by MAF, the Burden(1,25)
p-values weighted by each annotation, and a STAAR-B(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_STAAR_B_1_1}: a vector of STAAR-B(1,1) p-values,
including Burden(1,1) p-value weighted by MAF, the Burden(1,1)
p-values weighted by each annotation, and a STAAR-B(1,1)
p-value by aggregating these p-values using Cauchy method.

\code{weight_all_1}: a matrix of ancestry-specific weights across
\code{B} base tests for scenario 1 (if \code{find_weight} = TRUE).

\code{weight_all_2}: a matrix of ancestry-specific weights across
\code{B} base tests for scenario 2 (if \code{find_weight} = TRUE).

\code{results_weight}: a list of p-values weighted by
MAF, annotations, with aggregated components across \code{B} base tests. The p-values
from each base test are combined across weighting scenarios 1 and 2 using Cauchy method
(if \code{find_weight} = TRUE).

\code{results_weight1}: a list of p-values weighted by
MAF, annotations, with aggregated components across \code{B} base tests. The p-values
from each base test correspond to weighting scenario 1 (if \code{find_weight} = TRUE).

\code{results_weight2}: a list of p-values weighted by
MAF, annotations, with aggregated components across \code{B} base tests. The p-values
from each base test correspond to weighting scenario 2 (if \code{find_weight} = TRUE).
}
\description{
The \code{AI-STAAR_Binary_SPA} function takes in genotype, the object from fitting the null
model, and functional annotation data to analyze the association between a
imbalanced case-control phenotype and a variant-set by using an ancestry-informed STAAR-SPA procedure.
For each variant-set, the ancestry-informed STAAR-B p-value is a p-value from an omnibus test
that aggregated Burden(1,25) and Burden(1,1) together with p-values of each test weighted by each annotation
using Cauchy method, across an user-defined number of base tests. The p-values from each base test
are weighted by ancestry-specific ensemble weights estimated independently from the data.
}
\references{
Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}, \emph{52}(9), 969-983.
(\href{https://doi.org/10.1038/s41588-020-0676-4}{pub})

Li, Z., Li, X., et al. (2022). A framework for detecting
noncoding rare-variant associations of large-scale whole-genome sequencing
studies. \emph{Nature Methods}, \emph{19}(12), 1599-1611.
(\href{https://doi.org/10.1038/s41592-022-01640-x}{pub})

Liu, Y., et al. (2019). Acat: A fast and powerful p value combination
method for rare-variant analysis in sequencing studies.
\emph{The American Journal of Human Genetics}, \emph{104}(3), 410-421.
(\href{https://doi.org/10.1016/j.ajhg.2019.01.002}{pub})

Li, Z., Li, X., et al. (2020). Dynamic scan procedure for
detecting rare-variant association regions in whole-genome sequencing studies.
\emph{The American Journal of Human Genetics}, \emph{104}(5), 802-814.
(\href{https://doi.org/10.1016/j.ajhg.2019.03.002}{pub})
}
