% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCT.R
\name{CCT.pval}
\alias{CCT.pval}
\title{An analytical p-value combination method using the Cauchy distribution}
\usage{
CCT.pval(Pvals)
}
\arguments{
\item{Pvals}{a numeric vector of p-values, where each of the element is
between 0 to 1, to be combined.}
}
\value{
\code{pval}: the aggregated p-value combining p-values from the vector
\code{Pvals}.
}
\description{
The \code{CCT.pval} function takes in a vector of p-values, and return the
aggregated p-value using Cauchy method.
}
\examples{
p.values <- c(2e-02,4e-04,0.2,0.1,0.8)
CCT.pval(Pvals=p.values)
}
\references{
Liu, Y. and Xie, J. (2019). Cauchy combination test: a powerful test with analytic
p-value calculation under arbitrary dependency structures.
\emph{Journal of American Statistical Association (In press)}.
(\href{https://amstat.tandfonline.com/doi/abs/10.1080/01621459.2018.1554485?journalCode=uasa20}{pub})
}
