% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STAAR_cond.R
\name{STAAR_cond}
\alias{STAAR_cond}
\title{STAAR procedure for conditional analysis using omnibus test}
\usage{
STAAR_cond(
  genotype,
  genotype_adj,
  obj_nullmodel,
  annotation_phred = NULL,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants.}

\item{genotype_adj}{an n*p_adj genotype matrix (dosage matrix) of the target
sequence, where n is the sample size and p_adj is the number of genetic variants
to be adjusted for in conditional analysis (or a vector of a single variant with length n
if p_adj is 1).}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glm}} function for unrelated samples or
\code{\link{fit_null_glmmkin}} function for related samples. Note that \code{\link{fit_null_glmmkin}}
is a wrapper of the \code{\link{glmmkin}} function from the \code{\link{GMMAT}} package.}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p).
Continuous scores should be given in PHRED score scale, where the PHRED score
of j-th variant is defined to be -10*log10(rank(-score_j)/total) across the genome. (Binary)
categorical scores should be taking values 0 or 1, where 1 is functional and 0 is
non-functional. If not provided, STAAR will perform the
SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25), ACAT-V(1,1)
and ACAT-O tests (default = NULL).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default is 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default is 2).}
}
\value{
a list with the following members:

\code{num_variant}: the number of variants with minor allele frequency > 0 and less than
\code{rare_maf_cutoff} in the given variant-set that are used for performing the
variant-set using STAAR.

\code{RV_label}: the boolean vector indicating whether each variant in the given
variant-set has minor allele frequency > 0 and less than \code{rare_maf_cutoff}.

\code{results_STAAR_O_cond}: the conditional STAAR-O p-value that aggregated conditional
SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25), and ACAT-V(1,1) together
with conditional p-values of each test weighted by each annotation using Cauchy method.

\code{results_ACAT_O_cond}: the conditional ACAT-O p-value that aggregated conditional
SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25), and ACAT-V(1,1) using Cauchy method.

\code{results_STAAR_S_1_25_cond}: a vector of conditional STAAR-S(1,25) p-values,
including conditional SKAT(1,25) p-value weighted by MAF, the conditional SKAT(1,25)
p-values weighted by each annotation, and a conditional STAAR-S(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_STAAR_S_1_1_cond}: a vector of conditional STAAR-S(1,1) p-values,
including conditional SKAT(1,1) p-value weighted by MAF, the conditional SKAT(1,1)
p-values weighted by each annotation, and a conditional STAAR-S(1,1)
p-value by aggregating these p-values using Cauchy method.

\code{results_STAAR_B_1_25_cond}: a vector of conditional STAAR-B(1,25) p-values,
including conditional Burden(1,25) p-value weighted by MAF, the conditional Burden(1,25)
p-values weighted by each annotation, and a conditional STAAR-B(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_STAAR_B_1_1_cond}: a vector of conditional STAAR-B(1,1) p-values,
including conditional Burden(1,1) p-value weighted by MAF, the conditional Burden(1,1)
p-values weighted by each annotation, and a conditional STAAR-B(1,1)
p-value by aggregating these p-values using Cauchy method.

\code{results_STAAR_A_1_25_cond}: a vector of conditional STAAR-A(1,25) p-values,
including conditional ACAT-V(1,25) p-value weighted by MAF, the conditional ACAT-V(1,25)
p-values weighted by each annotation, and a conditional STAAR-A(1,25)
p-value by aggregating these p-values using Cauchy method.

\code{results_STAAR_A_1_1_cond}: a vector of conditional STAAR-A(1,1) p-values,
including conditional ACAT-V(1,1) p-value weighted by MAF, the conditional ACAT-V(1,1)
p-values weighted by each annotation, and a conditional STAAR-A(1,1)
p-value by aggregating these p-values using Cauchy method.
}
\description{
The \code{STAAR_cond} function takes in genotype, the genotype of variants to be
adjusted for in conditional analysis, the object from fitting the null
model, and functional annotation data to analyze the conditional association between a
quantitative/dichotomous phenotype and a variant-set by using STAAR procedure,
adjusting for a given list of variants. For each variant-set, the conditional
STAAR-O p-value is a p-value from an omnibus test that aggregated conditional
SKAT(1,25), SKAT(1,1), Burden(1,25), Burden(1,1), ACAT-V(1,25), and ACAT-V(1,1)
together with conditional p-values of each test weighted by each annotation
using Cauchy method.
}
\references{
Li, X., Li, Z. et al. (2020). Dynamic incorporation of multiple
in silico functional annotations empowers rare variant association analysis of
large whole-genome sequencing studies at scale. \emph{Nature Genetics}.
(\href{https://www.nature.com/articles/s41588-020-0676-4}{pub})

Liu, Y., et al. (2019). Acat: A fast and powerful p value combination
method for rare-variant analysis in sequencing studies.
\emph{The American Journal of Humann Genetics 104}(3), 410-421.
(\href{https://www.sciencedirect.com/science/article/pii/S0002929719300023}{pub})
}
