% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiv_Score_Test_Region.R
\name{Indiv_Score_Test_Region}
\alias{Indiv_Score_Test_Region}
\title{Score test for individual variants in a given variant-set}
\usage{
Indiv_Score_Test_Region(
  genotype,
  obj_nullmodel,
  rare_maf_cutoff = 0.01,
  rv_num_cutoff = 2
)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence,
where n is the sample size and p is the number of genetic variants.}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glm}} function for unrelated samples or
\code{\link{fit_null_glmmkin}} function for related samples. Note that \code{\link{fit_null_glmmkin}}
is a wrapper of the \code{\link{glmmkin}} function from the \code{\link{GMMAT}} package.}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.01).}

\item{rv_num_cutoff}{the cutoff of minimum number of variants of analyzing
a given variant-set (default = 2).}
}
\value{
A data frame with p rows corresponding to the p genetic variants in the given variant-set
and three columns: \code{Score} (the score test statistic), \code{SE} (the standard error associated
with the score test statistic), and \code{pvalue} (the score test p-value).
If a variant in the given variant-set has minor allele frequency = 0 or
greater than \code{rare_maf_cutoff}, the corresponding row will be \code{NA}. If a variant in
the given variant-set has standard error equal to 0, the p-value will be set as 1.
}
\description{
The \code{Indiv_Score_Test_Region} function takes in genotype and the object from fitting the null
model to analyze the associations between a quantitative/dichotomous phenotype and
all individual variants in a given variant-set by using score test.
}
\references{
Chen, H., et al. (2016). Control for population structure and relatedness for binary traits
in genetic association studies via logistic mixed models. \emph{The American Journal of Human Genetics}, \emph{98}(4), 653-666.
(\href{https://doi.org/10.1016/j.ajhg.2016.02.012}{pub})
}
